Refactoring) ExpDbSelectDialog file separated
